pyproject.toml
[project]
name = "check-datapackage"
version = "0.31.0"
description = "Check your Data Package's metadata against the Data Package standard to ensure that it's compliant with the standard"
authors = [
  {name = "Luke W. Johnston", email = "lwjohnst@gmail.com" },
  {name = "Kristiane Beicher", email = "kris.beicher@clin.au.dk" },
  {name = "Signe Kirk Brødbæk", email = "signekb@clin.au.dk" },
  {name = "Marton Vago", email = "marton.vago95@gmail.com" },
  {name = "Joel Ostblom", email = "joel@joelostblom.com" },
]
maintainers = [
  {name = "Luke W. Johnston", email = "lwjohnst@gmail.com" },
  {name = "Kristiane Beicher", email = "kris.beicher@clin.au.dk" },
  {name = "Signe Kirk Brødbæk", email = "signekb@clin.au.dk" },
  {name = "Marton Vago", email = "marton.vago95@gmail.com" },
  {name = "Joel Ostblom", email = "joel@joelostblom.com" },
]
readme = "README.md"
license = "MIT"
license-files = ["LICENSE.md"]
keywords = [
  "frictionless",
  "frictionless data",
  "verification",
  "metadata",
  "checks",
  "metadata verification",
  "data engineering",
  "metadata checks",
  "data package",
  "data package checks"
]
requires-python = ">=3.12"
dependencies = [
    "cyclopts>=3.13.5",
    "jsonschema>=4.25.1",
    "pydantic>=2.12.3",
    "python-jsonpath>=2.0.1",
    "rich>=14.3.0",
    "seedcase-flower>=0.3.0",
    "types-jsonschema>=4.25.1.20250822",
]

[project.scripts]
check-datapackage = "check_datapackage.cli:main"

[project.urls]
homepage = "https://check-datapackage.seedcase-project.org"
repository = "https://github.com/seedcase-project/check-datapackage"
changelog = "https://github.com/seedcase-project/check-datapackage/blob/main/CHANGELOG.md"
issues = "https://github.com/seedcase-project/check-datapackage/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "bandit>=1.8.6",
    "commitizen>=4.8.3",
    "genbadge>=1.1.2",
    "jupyter>=1.1.1",
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "quartodoc>=0.11.1",
    "ruff>=0.12.11",
    "types-tabulate>=0.9.0.20241207",
    "typos>=1.35.6",
    "vulture>=2.14",
]
